Update BlockNavigator.java
Added 2 methods at the bottom, took them from square driver and i use them to turn 90 degrees and to move forward 1 tile at a time.
Still needs the line detector and a way to get to the starting intersection. This class will only get you to the closest intersection to the point it needs to go to. I will put up a more complete copy in a few hours.